requirements.txt
streamlit==1.25.0
streamlit-drawable-canvas==0.8.0
torch
torchvision
pandas
sqlalchemy
psycopg2-binary
Pillow
